# PERF: lexsort_depth