--- a/pandas/_libs/algos.pyx
+++ b/pandas/_libs/algos.pyx
@@ -210,8 +210,7 @@ def lexsort_depth(list_of_arrays: list) -> Py_ssize_t:
 
     with nogil:
         depth = nlevels
-        i = 0
-        while i < n and depth != 0:
+        for i in range(1, n):
             k = 0
             while k < depth:
                 cur = vecs[k][i]
@@ -224,7 +223,9 @@ def lexsort_depth(list_of_arrays: list) -> Py_ssize_t:
                 else:
                     depth = min(k, depth)
                 k += 1
-            i += 1
+            if depth == 0:
+                # Depth can't increase, so if we've reached 0, break outer loop.
+                break
     free(vecs)
     return depth