@jbrockmendel 2022-06-26
Do we still need is_lexsorted?