@jreback 2022-07-03
does this help with any algos generally? e.g. try some indexing / multiindex benchs and see if any change